requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq>=0.9.0
crewai>=0.80.0
crewai-tools>=0.17.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=24.0.0

